changed the NTValues to allow double values.

+++ src/main/java/frc/robot/NTValue.java
@@ -7,16 +7,16 @@ import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 
 public class NTValue {
-  public NTValue (int initialValue, String key) {
+  public NTValue (double initialValue, String key) {
     ShuffleboardTab tab = Shuffleboard.getTab("NTValues");
     NetworkTableEntry entry = tab.add(key, initialValue)
       .withSize(2, 1)
       .withWidget(BuiltInWidgets.kTextView) // specify the widget here
       .getEntry();
     entry.addListener(event -> {
-      value = (int) entry.getValue().getDouble();
+      value = entry.getValue().getDouble();
     }, EntryListenerFlags.kNew | EntryListenerFlags.kUpdate);
     value = initialValue;
   }
-  public int value;
+  public double value;
 }
